recovery of old version without render graph.

+++ BlurOutline_PS.hlsl
@@ -16,7 +16,7 @@ cbuffer Control
 
 float4 main(float2 uv : Texcoord) : SV_Target
 {
-    uint width, height;
+    float width, height;
     tex.GetDimensions(width, height);
 
     float dx, dy;
